IVE changes for AKD support:
Force AKS arrays to scalarize the shape argument to 
vect2offset. This is done to permit index_optimize to
grab the name of the array B being indexed in B[iv].
Fear not, for the scalarized shape vector elements will 
be removed by DCR, if and when we happen to run it after
IVE completes.