Insert kludge code for -extrema support until -ssaiv works properly. The idea here is to attach AVIS_MAXVAL and AVIS_MINVAL values to WITHIDs, based on GENERATOR_BOUND1 and GENERATOR_BOUND2 values, but without -ssaiv, this is useless, because all partitions in a WL have the same WITHIDS, but different bounds.
Hence, we introduce, at the top of each WL partition, a temp
that gets the value of the WITHID, and then rename all the
code in the partition body to use the temp name instead of
the WITHID. At present, that is all that this phase does,
and it only does it if you compile with -extrema.

Oh, and at present, CVP is going to evaporate all my hard
work, but I'll fix that next.