Fixed bugs 473 and 537. UndoSSATransform now inserts assignments a=b to please markmemvals in precompile in two more situations: loops containing loops and conditionals containing loops.
I tried to eliminate all SSAASSIGN links during UndoSSATransform,
but that caused errors in the compilation of the stdlib. As it 
seems some phases make use of SSAASSIGN information even when the
ast is not in ssa form, which may cause trouble. Given the problems
I put my changes behind an #if for now.