Fixed bug 548.
Lacinlining used to downgrade AKV types to AKS when the formal loop function
parameter was only AKS, but the concrete argument was AKV, which is a very
common case for variables that represent loop carried dependencies.
However, this downgrade effects *all* applied occurrences of that variable
in the calling context. Due to common subexpression elimination these
can be numerous and entirely unrelated locations in the code.

We now do a proper downgrade of the loop variable's type.