Fixed bug 544.
This bug was actually due to a wrong handling of copied special functions
in inlining. Inlining was not properly adapted to run with grouped local
functions. Moreover, there was a conceptual problem with handling such
functions automatically in the traversal mechanism which, in conjunction
with the unusual and during the fun based cycle completely wrong traversal
order of inlining, caused the observed problems in dead function removal.
The reason for this was that copied special functions ended up in the
local function's chain of a function different from the one that actually
calls the special function. If now the regular function in which the special
function is wrongly attached to vanishes during dead function removal we
exactly get a broken AP_FUNDEF link.