This theoretically should not, hehehe, affect anyone's code, if they avoid -ecc, -check c,  -doawlf, and -extrema.
Basically, this is a half-done rewrite of IVEXP, changing
F_attachextrema into two primtives, F_noteminval and F_notemaxval.

I'm just checking it in because I have heard that a meteor is
going to hit rattler.snakeisland.com while I'm out for dinner,
and I don't want to lose a bunch of work.