This revision contains the mutc reference counting implementation with 3 modes norc, seq/c99m, async.
This revision changes the base type of the descriptor from int to
intptr_t so that pointers can also be stored in the descriptor.

New fields have been added to the descriptor to allow for different
reference counting modes.

This revision will be reverted and is here to allow more though
testing of these changes