Extend the type tuple: add bitarray flag
This flag will be used by the typesystem to detect whether a variable
corresponds to a bitarray or to a regular array. The possible values are YES,
NOT and ???. New macros SAC_ND_WRITE__BITARRAY and SAC_ND_READ__BITARRAY are
added but the code that tells the typesystem to substitute access to
bitarray-marked variables with those functions is commented out because it
breaks the compiler.