1. Rename sacprelude::Min          and sacprelude::Max to           sacprelude::intersectMin and sacprelude::intersectMax   so they can be used by AWLF.
2. Fix several AWLF bugs.

3. Tidy up cubeslicer a wee bit.

This version of AWLF appears to more or less work.
It suffers from the AL/AS/DL bug and the VPid bug
(aka IVE vs. LIR in Bodo's PM code), so we await
fixes for those. However, if you have an AKD problem 
you want to try it out on, do this:

  sac2c -O3 -doawlf foo.sac