Enhance CF to make ~/sac/testsuite/optimizations/constantfolding/SCCFprf_modarray11.sac work.
The problem here is that we have two concurrent modarray operations
on the same array element, e.g.,:

  a = [1,2,3,4];
  iv = id([2]);
  b = _modarray_AxVxS_( a, iv, 666);
  c = _modarray_AxVxS_( b, iv, 42);

We would like to replace the "b = ..." operation by

 b = a;

However, in order to do this, we have to preserve the
guards on iv. CF is looking at "c = ..." when it determines
that the replace is legal, but it is unable to
do the replace, due to traversal rules.

So, it makes the "b = ..." N_prf with PRF_ISNOP(b) = TRUE,
and a later CF traversal performs the desired replacement.