Fix bug whereby TC crashed after detecting type degredation: int{0} turning into int. This caused a number of CF unit tests to fail, when compiled with -doivecyc.
The type int{0} was determined in CYC, but was
lost when -doivecyc in SAACYC replaced an _sel_VxA_( iv, mat)
by an _vect2offset()/_idxsel() pair.
The direct problem was that COvect2offset was
not being invoked to simplify its constant arguments.

This fix corrects that problem. Unfortunately,
it uncovers another problem with vect2offset, 
namely an irrecoverable loss of index vector
information. For example, this CF unit test
now crashes in CF:

In SCCFprf_selivecyc.sac, we have:

 iv = [1, 1];
 XX  = [[two, three],[four, five]];
 z = _sel_VxA_(iv, XX);

CF should partially evaluate the selection, giving:

 z = _sel_VxA_( [1], [four, five]);

However, by the time that happens (running with -nocyc),
COvect2offset has done its damage, and we have:

  _ivesplit_76 = 3; 
  z = _idx_sel_( _ivesplit_76, XX); 

CF dies with the failing assert: "shape(iv) < dim(x)".

I do not see any way to deal with this in the current
code. What we can do is somehow ensure that vect2offset
preserves iv, perhaps by passing it as an additional result.
I.e.:

  offset, iv' = _vect2offset_( shp, iv);

Ideas welcome.