The commit fixes the bug 859.
After the investigation, it turned out that for the successful
loop unrolling we have to have `etc', and `vp' going in cyc
before lur.  Currently I just doubled this optimisations and
from saacyc and named them xetc and xvp.  This is *not* the 
best solution, but it works so far.

Now, some premature extrema work is done, but the code is
commented out, as it expects `id' in AVIS_{MIN,MAX} and 
I have an expression.  We can try either to create and id for
AVIS_{MIN,MAX} or to allow them store an expr.