DLIR: refactor for the new phase machanism
I do not understand the meaning of INFO_TRAVSTART in this traversal.
It looks to me that the DLIR does different things to a normal fundef when
the fundef is (a) reached from a module spine (the DLIR is invoked on
a module), or (b) the DLIR is directly invoked on that normal fundef.
Whether this is a desirable behaviour or a bug I cannot tell.
And, to complicate the matters more, the DLIR may be also invoked directly
on lacfuns in the AWLFI's SimplifySymbioticExpression().