WLF assumes that, if the result of a producerWL is AKS, then GENARRAY_SHAPE of that WL is AKV.
A recent extension to WLPROP, allowing it to work on CONDFUNS
(not checked in yet), causes an assert failure in WLF, because the 
producerWL, when copied into the CONDFUN, has AKS result
shape, but GENARRAY_SHAPE, now an N_arg to the CONDFUN,
is no longer AKV, because WLBSC got its fingers in the pie,
and generates an AKS N_array for that shape.

The change here merely makes the GENARRAY_SHAPE check for AKV
early enough that WLF does not attempt to perform the fold.

Presumably, on another trip through CYC/SAACYC, things
will untangle and the WLF will happen.