Merge of nested branch into trunk
This is way less clean and nice than I intented but this is mainly because the
original branch was only partly and screwed up by svn. Lets hope this one is
successfull.

Unit tests will be added to sac/sac repo later

The implentation chosen is one with a special nested keyword.

nested int[.] row

This creates two functions. enclose_row and disclose_row. Which are used for
enclosing and disclosing of one type to another. Note that selection does not
work by default and one should create its own select functions! (But this might
be automated lated).