I am checking this in only because of the pending repository move.
The idea behind this change is make SISI (which has been disabled
for Many Moons, anyway) operate ONLY on lacfuns, and to make
use of an LFU utility function for adding nodes to lacfuns
in a cleaner, or at least consistent, way.

If someone can present reasons why the old decision choice
mechanism in SISI is preferable to "lacfuns only", please feel
free to change it or back off this revision.