Rough cut at fixing Revision number for "sac2c -V" that were broken in the move from SVN to GIT.
The code now works by counting the number of commits, and
using that as a crude revision number. A monotonically increasing
revision number of some sort of required by any code that
wishes to automatically compare various revisions of
sac2c.

The code does NOT yet provide any support for counting the number
of local changes since commit to the head end. However, it is,
I think, an improvement over nothing, which is what we have now.