Introduce support for conformining extension to definition of _mod_(), following ISO APL Standard N8485 definition of residue:
1. If PRF_ARG2 is 0, result is PRF_ARG1, perhaps extended to match shape
   of PRF_ARG2.

2. Fix missing code in ct_prf.c that caused reference to undefined
   variable. This was detected by the compiler, but the code author
   did not see fit to correct the problem.

3. If PRF_ARG1 is negative, the result is positive, and suitable for
   use by rotate() in the other direction....

The definition of mod( ⍵, ⍺) in SAC now reflects the APL definition, except that
the argument order is reversed.

FYI: the definition, as presented
below, can be extended to real numbers and complex numbers without change.
In order to get the exact behavior of SHARP APL & ISO Standard APL on
those numbers, it would be necessary to introduce tolerant comparison
into the mix, so that floor (⌊) would produce the correct result.
This could probably best be achieved with a variant of mod() that
takes three arguments:

  mod( ⍵, ⍺, tolerance)

I do not propose to undertake that work, at least not until someone
shows a need for it.

------------------------------
Here is the definition for ⍺|⍵, taken from the
2nd Edition of the SHARP APL Reference Manual:

(⍺ is left argument, ⍵ is right argument)

  residue: ⍵-⍺×⌊⍵÷⍺+0=⍺

"A familiar use of residue is to determine the remainder
 resulting from dividing a non-negative integer by a positive
 integer... The definition {above} extends this notion to
 a zero left argument. When ⍺ is zero, the result is the
 right argument unchanged. It also extends the verb's
 domain to fractional right arguments and to negative
or fractional left arguments. When ⍺ is a negative integer,
the result ranges between ⍺ and zero, just as it does when
⍺ is positive."

The manual goes on to describe the tolerant behavior
of residue, a desirable property which arises from the
definition of floor, ceiling, and the relational functions
in APL. I am ignoring that for now, as well as ignoring
the behavior of residue on complex numbers, as
being out of scope for this discussion. If people are
interested, I'll provide that text. The residue
verb presented above is defined exactly the same on
all of those. The tolerance arises from the definition
of floor on real and complex numbers.