Fix Bug #1067.
Basically, this fix restricts the domain of AWLF on naked consumer-WLs
to _sel_VxA_( constant, PWL). Unfortunately, this means that some simple
code won't fold as desired, e.g.,:

    PWL = with(...);
    iv = shape(PWL)-1;
    z = _sel_VxA_( iv, PWL)

There may be a way to lift this restriction, but I don't see one
just now. The problem is that we have to determine which partition
of PWL intersects with iv, and we have no way to ensure that
iv won't change (as it would if it were loop-carried in a FOR-loop).
Waiter: More coffee, please!