replaced MEMmalloc framework
-d memcheck now creates a <filename>.mreport
this report contains information about the memory management of the phases.
this report is sorted on the number of space leaks per phase.

2 kinds of leak are possible:
1. The phase lost the reference to it
2. The malloc isn't freed before the end of the program

There are probably a lot of false positives because the check_attribs.c does not
correctly traverses every structure in the tree (This can be solved by
implementing the touch functions correctly)