Replacing (error|warn)Line to (error|warn)Loc.
After introducing proper locations in the compiler we need to adjust
existing calls to functions that produce errors or warnings to use
locations instead of lines.  The given set of files was easy to change,
in the other places error/warning message depends on global.linenum
which is being set automatically while traversing into the node.
This bit requires extra care.  I would propse to get rid of
global.linenum at all and use NODE_LOCATION on the according node.
Some discussion is required.