Disable NEEDCOUNT != WL_NEEDCOUNT check in AWLF, at least for now.
It will be interesting to see who complains about degraded AWLF
performance, as a result of this. Perhaps one The Annointed Ones
can point me at a failing example in real code.

I plan to write a cost function for PWL partitions, and let
that control the folding, probably this way:

 - If the "cost" of generating a PWL partition element is
   under some hand-waved threshold, allow the AWLF to occur.

 - Otherwise, allow the AWLF to occur if the NEEDCOUNTs
   show that there is only one CWL using the PWL elements.

This caused AWLF unit test bug652.sac to fail.
With this change, it runs clean.