Introduce CF extensions for eq() and neq(), when extrema are present on PRF_ARG1 or PRF_ARG2.
The motivation for this was the wlcondo.sac ( a relaxation
on vectors) constant-folding unit test,
created by rbe and sbs last weekend. Our goal
was to make this APLish code:

 mask = (iota(N+1) == 0) | (iota(N+1) == N);
 a = tod( mask) * a + tod( !mask) * ((shift([1], a) + shift([-1], a))/div);

generate the same IL as this one:

   a = with {
          ([0] <= [i] <= [0]) :  a[i];
          ([1] <= [i] <  [N]) :  ( a[i-1] + a[i+1]) / div;
          ([N] <= [i] <= [N]) :  a[i];
        } : modarray( a);

which it now does.

The CF unit tests run clean with this commit.

NB. There is still some chance of weirdness in CF on eq/neq, because
the modifications to code there were fairly substantial.