1. Various cleanup of AWLFI. The take([-1], PWL) code now folds    as desired. (AWLF unit test realrelaxAKDLowerA.sac)
2. IVEXP now refuses to propagate anything that has extrema
   from outside the current block.

3. AWLFI now fakes up extrema for IV in sel( IV, PWL) for those
   elements of IV that do not have extrema.

Items 2 and 3 make item 1 work.

Naked-consumer selects, such as those in
seliWL*sac, no longer fold, due to this change, and I am
not sure how to make them do so. The problem is that IV
has extrema from the outer WL, and they are too big for
the inner WL. It may be that we can handle this as follows:

If AVIS_DEFDEPTH( IV) == AVIS_DEFDEPTH( PWLresult), then
this is a naked consumer. If so, then we can ignore
any extrema on IV.