Enhance IVEXP propagation of extrema for _non_neg_val(). It will now overwrite the result AVIS_MIN, if it can show that the bounds are tightened as a result of that.
This is a key to making AWLF work on shift() properly.
The AWLF shifttest2.sac unit test now folds properly
on positive and negative, constant and hidden counts.

There are several other pieces of code that have to
be checked in here, including stdlib ArrayTransform.xsac
and the AWLF unit tests, so don't assume it's all working
yet.