Add new traversal, SAFA. This converts calls to commutative dyadic scalar functions so that they are in a canonical order.
The idea is to map expressions such as:

   z1 = _max_SxS_( x, y);
   z2 = _max_SxS_( y, x);

into a canonical form, so that CSE/CF can simplify them. In this
case, we would end up with:

   z1 = _max_SxS_( x, y);
   z2 = _max_SxS_( x, y);

and then CSE would give us:

   z1 = _max_SxS_( x, y);
   z2 = z1;

This arose in an inspection of a CF unit test failure in bug1082.sac.

In other news, the files changed here serve as a good example
of the minimum set of changes required to introduce a new
compiler traversal.