Extend ALisAssociativeAndCommutativePrf() to include eq/neq.
This is not required by AL, but is used by another caller of
that predicate function.  I have not made ANY changes to AL
other than this, so it may break in unforseen circumstances, if
eq/neq sneak into its realm. However, no failures of that
nature have appeared in either the CF nor AWLF unit test suites.