Fix a missing declaration bug.
Prior to this fix "config.h" is included after "unistd.h". However,
some declarations in unistd.h are dependent on configuration,
as the choice of source compatibility (__EXTENSIONS__, _POSIX_C_SOURCE
etc) may be sotred in there.

In particular, if _DARWIN_SOURCE is not defined on OSX, mkdtemp is not
declared. So at first use its return type defaults to int (instead
of char *). At run-time the int is casted to char*, dropping the high
order bits on platforms where sizeof(void*)>sizeof(int), and the pointer
becomes invalid. A segmentation violation ensues.

This patch avoids this entire mess by ensuring that config.h is included
before system headers.