Enhance the top-level "clean" make rule.
At the end of "make clean" an attempt is made to remove
all the remaining contents of bin/ and lib/. However on OSX
there may appear directories of the form XXXX.dSYM as soon
as a debugger has been used (eg gdb, valgrind). A simple
"rm" command fails to remove these and causes an error
in "make clean". Use "-rf" instead.