Repurpose SAC_MT_HM_AUX_THREADS for runtime specialization.
As discussed with Clemens, "AUX_THREADS" is a bit too general for
now, especially since nothing uses it yet. Since rtspec is going
to be the only user for now, be clearer in the name. If at some
point more auxillary threads are present in the system the naming
can be revisited.