Flush the SYScall tracking script before each command.
If a command run via SYScall takes too long it may happen that the user
kills both the command and the sac2c process. When this happens, prior
to this patch the last outputs to the script may still reside in the
process memory and thus be lost for inspection. By flushing before every
command we ensure that the last launched command line is always visible
for inspection.