Place the configure script at the top of the source tree.
This is important to limit user confusion. In a use case as
follows:

   cd sac2c
   ./configure --prefix=../inst

The intention of the user is to set the prefix to sac2c/../inst.
However prior to this patch "configure" is really run from sac2c/setup,
so the argument really sets the prefix to sac2c/setup/../inst, which is
confusing.

This patch fixes that issue by making sure the real "configure" script
resides at the top of the source tree. The other aspects of the
configuration remain unchanged.