Attempt to fix, in the absence of any high-level documentation about rcopt.c's goals in life, Bug #1107.
The code died when it came across an F_suballoc when it
expected an F_alloc. I inserted a case statement to handle
both of these, and crash if it encounters something else.

I have no idea if this fix is correct, and I also have no
idea why this code was not failing more often.