Parser location fixes part II.
Trying to insert sensible location for every call to TBmake<xxx>.
It seem to work, however the overall situation is error-prone.  I would
think, that TBmake<xxx> should include obligatory location parameter
when creating a code for nodes like:
    TBmakeNum, TBmakeId, etc.

and figure out location automatically for the nodes like:
    TBmakeExprs, TBmakeLet, etc.

Currently, the TCmake<xxx> would not propagate any location information,
so whenever we call them we risk to introduce wrong locations.

Also, there is a number of files which call CTIerrorLine instead of
CTIerrorLoc, as they rely on global.filenam and global.linenum that
is set by traversals.