Support mt_pth target when CC requires -pthread.
Here is a problem: when ax_pthread.m4 checks for pthread backend, in
case it is -pthread it puts it in the PTHREAD_CFLAGS, not in
PTHREAD_LIBS, which caused failures when compiling with -target mt_pth.

According to
https://lists.gnu.org/archive/html/autoconf-archive-maintainers/2013-06/msg00021.html
this is a normal behaviour and one should include what appears to be
CFLAGS in the linker call.

This patch does the following things:
    1. Include PTHREAD_CFLAGS into the linker call
    2. Renames pthread-related variables from CCMTXXX into CCMT_PTH_XXX
    3. Removes pthread-related flags from the mt target and moves it
       into mt_pth target.  Non all the mt backends need to link with
       pthreads, and if they do, they should ask for it explicitely.