Make glibc 2.20 shut up about deprecated _BSD_SOURCE.
Brilliant minds of glibc decided to deprecate _BSD_SOURCE and
_SVID_SOURCE by introducing _DEFAULT_SOURCE which is equivalent to:

    -D_BSD_SOURCE -D_SVID_SOURCE -D_POSIX_C_SOURCE=200809

This is cool of course, but only if you have the new version.  Old glibc
will complain.  So, the proper solution is to check the glibc version in
the configure file and to decide which defines to include.  The
intermediate solution is to define _DEFAULT_SOURCE *before* _BSD_SORUCE
and _SVID_SOURCE and forget about it.

Signed-off-by: Hans-Nikolai Viessmann <hansthespass@googlemail.com>