First cut at an ISL Polyhedral solver interface.
There is a definite problem with the build, because the Makefile has
to point to the barvinok source directory - there are a few things missing
from /usr/local/isl and /usr/local/include/isl.
Ergo, the bolted in pathname in the Makefile. I will fix this when I
hear back from the ISL people.

The build problem should not be serious at present, because the
tools/cuda Makefile is never executed unless you go there and invoke
it manually. This has to be repaired, but I do not know how to do that.