1. Introduce ISL interface for POGO,PHUT,PWLF.
2. Introduce new TU utility for Boolean coercion.

3. Desupport ALL PolyLib code in POGO,PHUT,PWLF.

There are two reasons for changes (1) and (3). First PolyLib
really does not let us support STEP and WIDTH. Second, in:

https://groups.yahoo.com/neo/groups/polylib/conversations/topics/105

Sven Verdoolaege says:

 { 2i | 0 <= i <= 10 } \ ({ 2i | 0 <= i <= 10 } union { 2i+1 | 0 <= i <= 10 })

The result should obviously be the empty set.
At some point during the computation polylib would compute

{ 2i | 0 <= i <= 10 } \ { 2i | 1 <= i <= 10 } \ { 0 }

(again, the result should be the empty set).

Thus, there is a bug in PolyLib that was never fixed, apparently, so
Sven recommends not using PolyLib at all, and to use, naturlich, ISL.
Which is where this commit takes us.