This claims to be the first working version of POGO/PHUT using ISL.
Both the pogo and pogorelational unit tests run clean, except for
a few problems (Extrema*sac) that will go away when PWLF works,
a crash in WLT, and a few other unresolved bugs that appear to be
unrelated to POGO.

Performance is fairly good. E.g, ipbb.sac -check c leaves 12
leftover guards with -nopogo, but zero with -dopogo.