Add support for different modes to rtspec.
Split the current implementation of controller & queue apart, so
we can have multiple different implementations concurrently available
for the user to choose from and to compare against.

Currently there is 3 modes: simple, hash and uuid.

simple corresponds to the initial implementation done by tvd.
uuid is the new implementation of a different dispatching mechanism
and function identification using UUIDs.
hash is a simplified implementation of uuid, where we use a hash
of various environment parameters + function information
for function identification.