Fix configure check for -Wstrict-prototypes.
We have agreed that function prototypes in the codebase must be strict.
I.e. a function with no arguments must be declared as

        int f(void)

and not as

        int f()

as in the latter case it is possible to call f with arbitrary number of
arguments.

The problem was that the check in configure that was verifying if CC
allows -Wstrict-prototypes flag was broken.  Now it is fixed and some
obvious warnings that showed up are fixed as well.  However, you might see
more warnings in case you compile some esoteric parts of the compiler.
Please do not complain about new warnings, just fix them instead.

Usually the fix is to replace declarations like "int f()" into "int f(void)".