(re-)introduced the MODIFIED tag for revision numbers and enabled non-installed builds
The goal of this work is to enable compiler developments in the source tree *without* installing not yet tested stuff and without requiring the use of a wrapper or similar.
We achieve this by two measures:
1) The file .revision.txt that is generated through check.mkf now adds ":MODIFIED" whenever the current branch has local modifications.
2) When the library loader (in sactools.h) neither succeeds through the LD_LIBRARY_PATH, nor through the install directory, it produces a warning and checks
     inside the build directory.
Jointly, these two measures enable our goal. Any local build that has not been installed will use the local stuff (measure 2).
Measure 1 ensures that even after an install the local stuff is used as soon as some sources have been modified.