hopefully completed the support for non-installed compiler runs
Here, we add sac2crc.local in setup which adds search paths as needed when the install directories are not present.
When sac2crc is not found in the install directory this sac2crc.local file is read in addition to the sac2crc in setup!

Also aded some more debug output and adjusted to the upcoming release in -V ;-)