Fix the dependency tracking.
Prior to this fix the list of dependency files (.d) was generated from
the list of per-target directories and files, including those
containing only headers. Since the rule to build a .d file only works
with a .c requirement, this would cause an error building the runtime
target.

This patch excludes all directories with suffix _h from the dependency
list.