Fix the build of the poly and ISL stuff.
Prior to this patch the ISL and polyhedral tools had their own
build rules in build.mkf, which did not adopt the same CFLAGS
and LDFLAGS settings as the rest of the sac2c build, and thus
yielding errors when the build is dependent on these vars being
set properly.

With this patch the regular build rules are applied to
these tools.