Fixed sac4c - wrong save location for c-library
Previously, sac4c would compile and link the C wrapper library to '.',
but due to the introduction of Targets/SBIs, there is now an expected
file directory hierarchy of the form: <path>/host/<target>.

The generated linker flags were updated to reflect this new hierarchy,
and so did not search in the right directories for the C wrapper library.

All sac4c generated C libraries will be stored in directory hierarchy of
the form: <path>/[cwrapper]/host/<target> where [cwrapper] can be set
with the -o flag.