Added traversal for marking functions for inlining
This traversal collects a few metrics, which based upon a simple
heruistic, determines which functions are good canidates for inlining
and marks them so. The actual inlining is achieved through the other
existing traversals, e.g. INL, PINL, etc.

NOTE: The traversal revealed a problem with classes, whereby inlining
causes 'information' from the AST to be lost, leading to compiler
errors during the code generation phase. As such, this traversal is
by default OFF. See bug #1172 for more details and hopefully a resolution.