Avoid polluting the library search path when compiling standalone programs.
Prior to this patch, if compiling a standalone program with -o the name
provided after -o was added as a library search path for the linker.
This is unnecessary and even yields an incorrect rpath.
This patch ensures that the directory passed via -o is only added
to the search path when compiling SAC modules or classes.