Break the circular dependency between libsac and libsacphm.
Prior to this patch, the error/warning functions would call SAC_HM_ShowDiagnostics, defined
in libsacphm, but the latter would in turn call functions in libsac. Some linkers do not
like this too much. To solve the problem this patch breaks the circular dependency,
by introducing a callback pointer in libsac that is initialized in libsacphm.