Don't mark every wrapper as indirect wrapper when in runtime mode.
I suppose this was done as an optimization step to not have to
re-mark the wrapper we are interested in as indirect when generating
specializations at runtime (the original info is generated after
serializing the syntax tree so no longer/not yet available when
loading the Tree file for the module we are specializing for).

However, we now load far more than just one particular wrapper, so
the optimization doesn't work anymore and actually causes bugs.