sac4c - fixed missing cwrapper libpath in -ldflags
Using sac4c with flag -ldflags generates a list of -L and -l flags.
Up until the full implementation of targets and SBIs, -L. was sufficent
to locate the libcwrapper.so file generated by sac4c in the local directory.
Now however, libcwrapper.so is located under ./cwrapper/host/SBI, and the
mechanism used to generated the string for -ldflags was not updated to reflect
this. This commit fixes this, and makes use of the global.outfilename variable
to specify the path.