Fixed silly DMG install script
Basically, cp on fApple (BSD-like-ish thingy) is sensitive to whether
or not there is a '/' at the end of the SRC parameter - if there is one,
do not include the basename dir of SRC in the copy action, otherwise do.

The result in our case was that we did *not* have the '/', and so copied
both the content of the basename dir and it as well to /usr/local (or
whatever the install prefix is).