Potential Solution to RTSPEC component compilation
In essence, the build system sets the some bars which indicate
whether or not uuid.h and/or crypt.h are avaible. These two headers
(so far) are only used in one place within the sac2c source code,
specifically within RTSPEC. If one of these is missing from the system,
then RTSPEC is turned off - however the variables persist and make it
into the config.h header - resulting in segments of RTSPEC being
included/compiled resulting in additonal symbols within libsac2c.so.
This in of itself isn't a problem, but because RTSPEC is off, we don't
propegate the relevent headers and libs, meaning that some of the
symbols are undefined.