Added automatic re-configuration to CMake build
We now track the state of the git repo, and on any change cause
a re-configuration to occur. The main advantage with this is that
the version of the compiler is propagated correctly to the build
system.

This commit relates to commit d2e5b769461a58d0, which introduces
a check to ensure that we do not install on a dirty build.