Fixed RPM/DEB packaging directory conflict
In essence, we were specifying an absolute path in the final
install macro. This lead CPack to believe that everythin under that
path must be included as well - which is wrong. Have changed the
destination path to a relative one.

It is not yet clear if this is the ideal solution.