Added binary version checking
The idea behind this change is to make looking for the libsac2c
library more smart. For instance, the current scheme of the build
system prevents installs of any non-commited changes. One problem
with this approach is that there is a chance that a local binary may
continue to try to load the globally installed libsac2c, and not the
one in the build directory. Using this version check solves this as
we can see immediatly if we are 'dirty' or not.

The actualy smartness will come in a next commit - stay tuned :-D