Made library searching/opening smarter
This relates to commit 197cf7cccf3.

In essence, when we are dirty we only try to load the local build
of the libsac2c. Otherwise we try to load the global one, and if
this fails we fallback on the local build.

Versioning ensures that we don't use the wrong library in either
context.