Correctly find sac2crc file now
This is a rather large commit. The main intent is to make the location
of sac2crc definable. This is mainly relevent for relocatable installs
with the ARCHIVE packages, such that a user can place the SAC libraries
and configure files anywhere and still access these correctly through
the binaries.

One concern is how we pass the sac2crc location to the libsac, perhaps
there is a more appropriate way of doing this.