We now generated archive install script
Based upon the code review, it was decided that most variables, such
as the postfix and sac2c version can be included into the install
script by the build system.

Additionally, the install script has had a bit of bug fixing done.
In particular the code has been refactored and simplified,
additionally checks on the command line parameters have been added,
and description of been expanded.