Omnibus packaging extended
As is the case with individual packaging, we can no create archives
and distributable packages that have different content depending on
their use case - archives contain the binary sources from which the
end-user can compile the binaries to have different search paths.

This is facilitated by exposing the components layout to the omnibus
packaging making facility.