Removed static global in resources.c
The global_t type has been extended to have a notion of global
variables that are not initialised (as is done in GLOBinitializeGlobal).
The purpose of this is to allow for global variables to be set
prior to any other global variables being initialised.

This is the case with the sac2crc search paths, which must be set
before the function in main.c are called. This replaces the formal
solution that used `static global` with resource.c to achieve the
same effect.

This commit is based on 8af123c3f14ace02.