Added archive installer wrapper
In essence, for the omnibus package (which contains all release
types of the compiler), I have created a wrapper script around the
actual install scripts for each release type.

There is some room for improvement in this design, but given that
we are getting this `for free' by CMake, I don't think its a large
maintenance overhead.