Fixed wrong working directory on GIT commands
We use GIT for two things,
 1. to get the version name of sac2c (using `git describe`)
 2. to check the dirty state of the repository

In both cases, the initial implementation did not specify the working
directory of GIT. Most of the testing of this was down within the
directory call `build' under the source tree, and as such there were
no problems. For users who prefer to build in `/tmp`, this lead
to GIT not finding the repository, and so these commands failed
leading to **strange** behavour in build system.

Additionally, this commit removes the CHECK_IS_DIRTY macro and
inlines it instead, this simplifies things for the user.