Moved sac2crc stuff from runtime up to sac2c build
Before this commit, we were generating two sac2crc.local files, one
in the build dir of sac2c and the other in the build dir of the runtime.
The runtime sac2crc.local howver is copied up and overwrites the other
sac2crc.local file. This was done because it wasn't clear how to access
some CMAKE generated variables within the runtime build outwith.

Additionally, this commit introduces a prelude sac2crc file, this serves
no local purpose but is meant to be copied into the user's .sac2crc
directory in their home dir.