Resolve CPATH and LD_LIBRARY_PATH to find includes and libs
Cmake does not by defualt use these environment variables to find includes
and libs that are not within standard search paths. Cmake does however
provide internal varables that can be set to look for additional paths.

We now search in the paths as given by CPATH and LD_LIBRARY_PATH.