Apply Bandaid onto RPM generation
In essence, RPM generation works fine, except that it assumes all
libraries for
SaC are located under the `lib' directory, which for the runtime
libraries
is not the case (which are under lib/host/.../... etc.). As such,
when resolving dependencies, the RPM generator sees that sac2c depends
on libsac, but misses out that SaC provides libsac. When we reach
the stage of installing (via `yum install ...rpm'), we get errors
telling us that dependencies can't be matched as no packages for
libsac can be found.

This patch is a 'bandaid' because it is not a -good- fix - I have
verbatim updated the RPM provides field. Ideally we should
auto-generate
this, but that would require several changes - importantly passing on
the location of the libs during configure time *and* analysing them -
which isn't necessarily possible.