Fix potential for overflowing syscall buffer
For doing syscall, we did use a static char* buffer set to a fixed
size of MAX_SYSCALL (or 10000 elements). When SaC was young, this did
not pose problems as typically one would never make a syscall with
more characters that this. Of late, we are passing more arguments
and using absolute paths, that we now have reached this limit.

This patch moves away from char* buffers, and to using the str_buf
(SBUF) data structure instead, which supports dynamic resizing. All
syscall functions have been updated to use this.