Fix POGO treatment of zero-trip WL partition on non_neg_val(iv)
Fix bug that happens when PHUT/POGO encounter
a zero-trip WL partition on non_neg_val(iv). That was causing a
ISL union set to be a null set. It should return TRUE,
as an empty set contains no negative values.

There are likely similar problems with the other guards, but I have
not looked into this area yet.