Fix ipbb.sac crash & simpleConstantsUp.sac infinite loop problem
POGO would like to know information about loop counts
in loopfuns, EXCEPT when the POGOprf is the COND_COND
that controls looping.

We now generate the AFT for the loopfun in POGOfundef,
and then pass the AFT to ISL, to be intersected with exprsX
before proceeding. In the "except" case above, we pass
NULL, thereby preventing the COND_CON from being replaced
by TRUE.