Fix PHUT predicate fns' treatment of ISL set variables
Fix PHUTisPositive, etc.
I am not sure if this fix is the right way to go or not.
Basically, I have an AFT built for (in this case) POGO,
and I need to know if one of the arguments is positive, negative, etc.
The above predicate functions invoke set intersection on
the argument's AFT, but ISL requires that the PHUTisPositive
argument be an ISL set variable.

What we do now is to pass the variable name to the intersect
code, and temporarily make it a set variable. If there is
a cleaner way to do this, I am all for it.

At any rate, it cleaned up one of the ipbb.sac unit test failures.
The unit test still fails, but for unrelated reasons.