Introduce PHUT AFT search function for loop-carried variable
that is an argument to a function whose result appears in condprf.

This is needed for the PLUR/LoopFunNotEquals.doctz.sac unit test,
and other Loopfuns using -doctz. This alone is not enough to make
this unit test work, as I still need to make PHUR/PLUR turn the != into
<.